refactor: extract keepalive mixin from coordinator core
